vtfpp: ImageConversion::alphaToDistance -> DistanceMapping::alphaToDistance
