+++ include/vtfpp/vtfpp.h
@@ -5,6 +5,7 @@
  * include it the same way as any of the other SourcePP libraries.
  */
 
+#include "DistanceMapping.h"
 #include "HOT.h"
 #include "ImageConversion.h"
 #include "ImageFormats.h"
